Harden startup ordering, ACL/rate validation, and CLI error localization
Deep-scan rounds 11-12 fixes across daemon, store, web, and CLI:

- daemon: gate validateVPNListenAddr and the loopback probe behind the
  same hy2/REALITY predicate so pure-WSS multi-homed binds no longer
  fatal; serve the loopback WebSocket before REALITY/hy2 begin accepting
  (split reality listener bind from accept) to avoid startup-window
  bridge timeouts; defer the login success frame until VIP allocation
  and lease persistence succeed, sending an explicit failure frame on
  alloc failure.
- ACL: normalize per-rule and default actions fail-closed (unknown or
  non-canonical values become deny); make exit_dns_redirect deferred
  diff case/alias aware.
- store/config: align mesh_enabled and rate setting write validation
  with their read paths; add clamp-aware rate_burst_bytes validator with
  bounds shared as the single source of truth; reject negative durations.
- CLI: localize store.ErrNotFound for user/device lookups across exit,
  route, acl, device, credentials, and profile commands; canonicalize
  validated settings before persisting.
- web: enforce explicit flag > env > default precedence for all
  env-backed flags and cover it with tests; remove the dead -debug flag.

Adds regression tests throughout; go build, vet, and race suite pass.

+++ cmd/nanotund/rate_limit_test.go
@@ -175,11 +175,11 @@ func TestEffectiveBurst(t *testing.T) {
 		{"zero falls back to default", 0, defaultRateBurstBytes},
 		{"negative falls back to default", -1, defaultRateBurstBytes},
 		{"too small is clamped up", 1024, minRateBurstBytes},
-		{"min boundary unchanged", minRateBurstBytes, minRateBurstBytes},
+		{"min boundary unchanged", int64(minRateBurstBytes), minRateBurstBytes},
 		{"normal value passes through", 256 * 1024, 256 * 1024},
 		// N11:上界保护。
-		{"max boundary unchanged", maxRateBurstBytes, maxRateBurstBytes},
-		{"just over max clamped down", maxRateBurstBytes + 1, maxRateBurstBytes},
+		{"max boundary unchanged", int64(maxRateBurstBytes), maxRateBurstBytes},
+		{"just over max clamped down", int64(maxRateBurstBytes) + 1, maxRateBurstBytes},
 		{"1 GiB clamped to max", 1 << 30, maxRateBurstBytes},
 		{"int64 max clamped to max", 1 << 62, maxRateBurstBytes},
 	}
